FEV1232: Moved tap/scroll/find functions into separate script (#176)
* FEV1232: Moved tap/scroll/find functions into separate script

* changed commands.js to navigation.js

* moved functions for button/dropdown questions into controls.js

* replace triple/quadruple tap with multi_tap

* use deviceInfo as argument

* Move text entry logic to controls.js

* move radio/checkbox logic into controls.js

* updated comments
